persister/requirements.txt
kafka-python==2.2.2
nltk
pymongo